Casi acabado, solo falta detallar funciones y aplicar enum en todos los metodos

+++ src/mx/iteso/javaproyect/Matriznxn.java
@@ -5,7 +5,7 @@ public class Matriznxn extends MetodosMatriz{
 	
 	// determinante
 
-	public Matriznxn(double[][] MR) {
+	public Matriznxn(int[][] MR) {
 		super(MR);
 	}
