+++ src/mx/iteso/javaproyect/OperatorsMatriz.java
@@ -2,10 +2,10 @@ package mx.iteso.javaproyect;
 
 public interface OperatorsMatriz {
 
-	public abstract double[][] productoxEscalar(int z);
-	public abstract boolean baseCanonica(double[][] lolo);
-	public abstract double [][] suma(double[][] x);
-	public abstract double [][] resta(double[][] y);
-	double[][] MultipM(double[][] r);	
+	public abstract int[][] productoxEscalar(int z);
+	public abstract boolean baseCanonica(int[][] lolo);
+	public abstract int [][] suma(int[][] x);
+	public abstract int [][] resta(int[][] y);
+	int[][] MultipM(int[][] r);	
 	
 }
