+++ src/mx/iteso/javaproyect/OperatorsVect.java
@@ -3,11 +3,11 @@ package mx.iteso.javaproyect;
 public interface OperatorsVect  {
 
 
-	public double[] productoxEscalar(int y);
-	public double[] suma(double[] lolo);
-	public double[] resta(double[] lolo);
-	public boolean baseCanonica(double[] lolo);
-	public double[] productoCruz(double[] lolo);
+	public int[] productoxEscalar(int y);
+	public int[] suma(int[] lolo);
+	public int[] resta(int[] lolo);
+	public boolean baseCanonica(int[] lolo);
+	public int[] productoCruz(int[] lolo);
